refactor: move mod tasks to their own subfolder
Makes the launcher/minecraft/mod/ folder a little more organized.
