+++ launcher/minecraft/mod/tasks/ModFolderLoadTask.cpp
